requirements.txt
streamlit
plotly
pandas
numpy
pyarrow
scikit-learn>=0.24
openpyxl
statsmodels
prophet
xgboost
torch
pyyaml
mlflow
joblib
tqdm
optuna
pytest